pyproject.toml
[tool.poetry]
name = "resume-parser"
version = "0.1.0"
description = ""
authors = ["Mani Sankar Chintagunti <manisankarchintagunti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
groq = "^0.10.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
